Moved algorithm_specifiers.rs into utils
